Refactored API classes (ITSystemResource, HardwareResource).
