+++ oim_cms/test_api.py
@@ -226,19 +226,19 @@ class ITSystemResourceTestCase(ApiTestCase):
     def test_list(self):
         """Test the LocationResource list response
         """
-        url = '/api/itsystems'
+        url = '/api/itsystems/'
         response = self.client.get(url)
         self.assertEqual(response.status_code, 200)
         # The 'development' IT system won't be in the response.
         self.assertNotContains(response, self.it2.name)
         # Test all request parameter.
-        url = '/api/itsystems?all'
+        url = '/api/itsystems/?all'
         response = self.client.get(url)
         self.assertEqual(response.status_code, 200)
         # The 'development' IT system will be in the response.
         self.assertContains(response, self.it2.name)
         # Test filtering by system_id
-        url = '/api/itsystems?system_id={}'.format(self.it1.system_id)
+        url = '/api/itsystems/?system_id={}'.format(self.it1.system_id)
         response = self.client.get(url)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, self.it1.name)
